+++ http-pdf-generator/functions/src/parse_parameters.ts
@@ -1,13 +1,12 @@
 import { uuidv4 } from "@firebase/util";
 import { PDFOptions } from "puppeteer";
-import { ParsedQs } from "qs";
 import { extensionParameters } from "lib/utilities/extension_parameters";
-import { ParsedParameters } from "lib/utilities/parameters";
+import { ParsedParameters, TemplateParameters } from "lib/utilities/parameters";
 
-export interface GetParameters {
+export interface PostParameters {
   adjustHeightToFit?: "yes" | "no";
   chromiumPdfOptions?: PDFOptions;
-  data?: ParsedQs;
+  data?: TemplateParameters;
   headful?: "true" | "false";
   outputStorageBucket?: string;
   outputStoragePrefix?: string;
@@ -31,15 +30,15 @@ const BOOLEAN_PDF_OPTIONS = [
  * @return {ParsedParameters}
  */
 export function parseParameters({
-  getParameters,
+  postParameters,
 }: {
-  getParameters: GetParameters;
+  postParameters: PostParameters;
 }): ParsedParameters {
-  if (typeof getParameters.data === "string") {
+  if (typeof postParameters.data === "string") {
     throw new Error("'data' should be an object, not a string.");
   }
 
-  if (getParameters.data instanceof Array) {
+  if (postParameters.data instanceof Array) {
     throw new Error("'data' should be an object, not an array.");
   }
 
@@ -54,31 +53,37 @@ export function parseParameters({
   }
 
   if (
-    getParameters.templatePath != null &&
-    typeof getParameters.templatePath !== "string"
+    postParameters.templatePath != null &&
+    typeof postParameters.templatePath !== "string"
   ) {
     throw new Error("'templatePath' should be a string.");
   }
 
   const templatePath =
-    getParameters.templatePath ?? extensionParameters.TEMPLATE_PATH;
+    postParameters["templatePath"] ?? extensionParameters.TEMPLATE_PATH;
 
   const parts = templatePath.split("/");
   const templateBucket = parts[0];
   const templatePrefix =
     parts.length > 2 ? parts.slice(1, -1).join("/") + "/" : "";
+  console.log(
+    templatePath,
+    postParameters,
+    postParameters["templatePath"],
+    typeof postParameters
+  );
   const templateId = parts[parts.length - 1];
 
   const parameters = {
     adjustHeightToFit:
-      (getParameters.adjustHeightToFit?.toLowerCase() ??
+      (postParameters.adjustHeightToFit?.toLowerCase() ??
         extensionParameters.ADJUST_HEIGHT_TO_FIT?.toLowerCase()) === "yes",
     // Convert strings representing boolean values to booleans
     chromiumPdfOptions: Object.fromEntries(
       Object.entries(
         ({
           ...parsedChromiumPdfOptions,
-          ...getParameters.chromiumPdfOptions,
+          ...postParameters.chromiumPdfOptions,
         } as PDFOptions) ?? {}
       ).map(([key, value]) => {
         if (BOOLEAN_PDF_OPTIONS.includes(key)) {
@@ -87,29 +92,29 @@ export function parseParameters({
         return [key, value];
       })
     ),
-    data: getParameters.data ?? {},
+    data: postParameters.data ?? {},
     headless:
       process.env.IS_LOCAL == "true"
-        ? !(getParameters.headful?.toLowerCase() === "true")
+        ? !(postParameters.headful?.toLowerCase() === "true")
         : true,
     outputStorageBucket:
-      getParameters.outputStorageBucket ??
+      postParameters.outputStorageBucket ??
       extensionParameters.OUTPUT_STORAGE_BUCKET,
     outputStoragePrefix:
-      getParameters.outputStoragePrefix ??
+      postParameters.outputStoragePrefix ??
       extensionParameters.OUTPUT_STORAGE_PREFIX,
     outputFileName:
-      getParameters.outputFileName ??
+      postParameters.outputFileName ??
       `${uuidv4()}-${new Date().toISOString()}.pdf`,
     templateBucket,
     templatePrefix,
     templateId,
     networkIdleTime: Number.parseInt(
-      getParameters.networkIdleTime ??
+      postParameters.networkIdleTime ??
         (extensionParameters.NETWORK_IDLE_TIME || "0")
     ),
     shouldWaitForIsReady:
-      (getParameters.shouldWaitForIsReady?.toLowerCase() ??
+      (postParameters.shouldWaitForIsReady?.toLowerCase() ??
         extensionParameters.SHOULD_WAIT_FOR_IS_READY.toLowerCase()) === "yes",
   };